requirements.txt
# Core dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
httpx>=0.26.0
aiohttp>=3.9.0
tenacity>=8.2.3

# vLLM client (OpenAI-compatible)
openai>=1.12.0

# PDF processing
marker-pdf>=1.0.0
pymupdf4llm>=0.0.5
pymupdf>=1.24.0
pdfminer.six>=20231228

# Document processing
python-docx>=1.1.0
openpyxl>=3.1.2
beautifulsoup4>=4.12.3
lxml>=5.1.0

# Text processing
tiktoken>=0.5.2
nltk>=3.8.1
spacy>=3.7.2

# Embeddings and ML
numpy>=1.26.0
scipy>=1.12.0
scikit-learn>=1.4.0
sentence-transformers>=2.3.0  # Fallback for local embeddings

# Knowledge graph
networkx>=3.2
pyvis>=0.3.2

# Neo4j (optional persistence)
neo4j>=5.16.0

# Web scraping
scrapy>=2.11.0
requests>=2.31.0

# Google Forms integration
google-api-python-client>=2.116.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

# Logging and monitoring
loguru>=0.7.2
rich>=13.7.0
prometheus-client>=0.19.0

# CLI
typer>=0.9.0
click>=8.1.7

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Development
black>=24.1.0
isort>=5.13.0
mypy>=1.8.0
ruff>=0.1.14
